AS-3210: Fix column type in farms_webapp_pkg.insert_import_version.

+++ farms-liquibase/scripts/01_00_xx/01_00_03/00/ddl/routines/farms_webapp_pkg.insert_import_version.sql
@@ -1,5 +1,5 @@
 drop procedure if exists farms_webapp_pkg.insert_import_version(
-    numeric,
+    bigint,
     varchar,
     varchar,
     varchar,
